refactor: provider 타입 변경
