--- src/main/java/com/stcom/smartmealtable/domain/member/SocialLoginProvider.java
@@ -1,5 +1,0 @@
-package com.stcom.smartmealtable.domain.member;
-
-public enum SocialLoginProvider {
-    GOOGLE, KAKAO
-}
