+++ src/main/java/com/stcom/smartmealtable/domain/social/SocialAccountRepository.java
@@ -1,0 +1,9 @@
+package com.stcom.smartmealtable.domain.social;
+
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SocialAccountRepository extends JpaRepository<SocialAccount, Long> {
+
+    Optional<SocialAccount> findByProviderAndProviderUserId(String provider, String providerUserId);
+}
